+++ apps/user/src/components/password/passwordContent/passwordContent.hook.ts
